chore: move sync schema utils to separate sync.go to reduce the main driver file size
